+++ packages/transaction-pay-controller/src/strategy/fiat/fiat-submit-simple.ts
@@ -39,6 +39,10 @@ export async function submitSimpleRelay({
 
   const originalRelayQuote = request.quotes[0].original.relayQuote;
 
+  if (!originalRelayQuote) {
+    throw new Error('Missing Relay quote for fiat submission');
+  }
+
   const relayRequest: QuoteRequest = {
     ...baseRequest,
     isMaxAmount: false,